+++ myqueue/resubmit.py
@@ -9,16 +9,14 @@ from myqueue.submitting import submit
 def resubmit(queue: Queue,
              selection: Selection,
              resources: Resources | None,
-             force: bool = False) -> None:
+             remove: bool = False) -> None:
     """Resubmit tasks."""
     tasks = []
 
-    for task in selection.select(queue.tasks):
+    ids = []
+    for task in queue.select(selection):
         if task.state in {'queued', 'hold', 'running'}:
             continue
-
-        queue.tasks.remove(task)
-        queue.changed.add(task)
 
         task = Task(task.cmd,
                     deps=task.deps,
@@ -29,5 +27,9 @@ def resubmit(queue: Queue,
                     creates=task.creates,
                     diskspace=0)
         tasks.append(task)
+        ids.append(task.id)
 
-    submit(queue, tasks, force=force)
+    if remove and not queue.dry_run:
+        queue.remove(ids)
+
+    submit(queue, tasks)